requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

flask==2.2.2
Flask-Session==0.5.0
Werkzeug==2.2.3
boto3==1.28.80
mysql-connector-python==8.2.0
python-dotenv==1.0.0